+++ src/index.js
@@ -2,17 +2,13 @@ import './styles/style.css';
 import fetchAirPollutionData from './modules/airQualityData';
 import createCurrentWeather from './modules/currentWeather/createCurrentWeather';
 import getDates from './modules/forecast/chart';
+import toggleDarkMode from './modules/toggle';
+import loadContents from './modules/loadContents';
 
 window.onload = () => {
+  toggleDarkMode();
   fetchAirPollutionData();
   createCurrentWeather();
   getDates();
-  const displayContents = () => {
-    document.getElementById('loader').style.display = 'none';
-    document.getElementById('main-section').style.display = 'block';
-  };
-  const loadContents = () => {
-    setTimeout(displayContents, 5000);
-  };
   loadContents();
 };
