+++ src/styles/_toggle_button.scss
@@ -1,8 +1,10 @@
-label {
+button {
   width: 80px;
   height: 45px;
   position: relative;
-  display: block;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   background: #ebebeb;
   border-radius: 200px;
   box-shadow: inset 0 5px 15px rgba(0, 0, 0, 0.4), inset 0 -5px 15px rgba(255, 255, 255, 0.4);
@@ -10,32 +12,8 @@ label {
   transition: 0.3s;
 }
 
-label::after {
-  content: "";
-  width: 25px;
-  height: 25px;
+#toggle-icon {
   position: absolute;
-  top: 10px;
-  left: 10px;
-  background-image: url('../moon.png');
-  background-size: cover;
-  border-radius: 180px;
-  transition: 0.3s;
-}
-
-input[type=checkbox] {
-  width: 0;
-  height: 0;
-  visibility: hidden;
-}
-
-input:checked + label {
-  background: #242424;
-}
-
-input:checked + label::after {
-  left: 70px;
-  transform: translateX(-100%);
-  background-image: url('../sun.png');
-  background-size: cover;
+  width: 40%;
+  height: auto;
 }
