+++ src/styles/style.css
@@ -11,11 +11,19 @@ body {
   height: auto;
 }
 
-label {
+button {
   width: 80px;
   height: 45px;
   position: relative;
-  display: block;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
   background: #ebebeb;
   border-radius: 200px;
   -webkit-box-shadow: inset 0 5px 15px rgba(0, 0, 0, 0.4), inset 0 -5px 15px rgba(255, 255, 255, 0.4);
@@ -25,36 +33,10 @@ label {
   transition: 0.3s;
 }
 
-label::after {
-  content: "";
-  width: 25px;
-  height: 25px;
+#toggle-icon {
   position: absolute;
-  top: 10px;
-  left: 10px;
-  background-image: url("../moon.png");
-  background-size: cover;
-  border-radius: 180px;
-  -webkit-transition: 0.3s;
-  transition: 0.3s;
-}
-
-input[type=checkbox] {
-  width: 0;
-  height: 0;
-  visibility: hidden;
-}
-
-input:checked + label {
-  background: #242424;
-}
-
-input:checked + label::after {
-  left: 70px;
-  -webkit-transform: translateX(-100%);
-          transform: translateX(-100%);
-  background-image: url("../sun.png");
-  background-size: cover;
+  width: 40%;
+  height: auto;
 }
 
 .carousel {
@@ -132,34 +114,6 @@ input:checked + label::after {
     bottom: 0;
     opacity: 1;
   }
-}
-
-.mapControls {
-  margin-top: 10px;
-  border: 1px solid transparent;
-  border-radius: 2px 0 0 2px;
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  height: 32px;
-  outline: none;
-  -webkit-box-shadow: 0 2px 6px rgba(0, 0, 0, 0.3);
-          box-shadow: 0 2px 6px rgba(0, 0, 0, 0.3);
-}
-
-#searchMapInput {
-  background-color: #fff;
-  font-family: Roboto;
-  font-size: 15px;
-  font-weight: 300;
-  margin-left: 12px;
-  padding: 0 11px 0 13px;
-  text-overflow: ellipsis;
-  width: 50%;
-}
-
-#searchMapInput:focus {
-  border-color: #4d90fe;
 }
 
 #main-section {
